+++ src/php/npm-postinstall.php
@@ -1,5 +1,6 @@
 <?php
-if (is_readable(__DIR__ . '/../../gulpfile-config.dist.json') && is_writable(__DIR__ . '/../../')) {
+if (!file_exists(__DIR__ . '/../../gulpfile-config.json') 
+        && is_readable(__DIR__ . '/../../gulpfile-config.dist.json') 
+        && is_writable(__DIR__ . '/../../')) {
     copy(__DIR__ . '/../../gulpfile-config.dist.json', __DIR__ . '/../../gulpfile-config.json');
 }
-